// src/types/Client.ts
export type  Client = {
  _id: string
  name: string
  address: string
  postCode: string
  email: string
  nip: string
  regon: string
  founder: string
  founder2: string
  contractSignDate: string
  contractEndDate: string
  oneTimeFee: string
  oneTimeText: string
  subscription: string
  subscriptionText: string
  extraFee1: string
  phoneNumber: string
}